+++ codeInsight/impl/com/intellij/codeInsight/daemon/impl/quickfix/MakeMethodConstructorFix.java
@@ -2,6 +2,7 @@ package com.intellij.codeInsight.daemon.impl.quickfix;
 
 import com.intellij.codeInsight.daemon.QuickFixBundle;
 import com.intellij.codeInsight.intention.IntentionAction;
+import com.intellij.codeInsight.CodeInsightUtilBase;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiFile;
@@ -34,11 +35,8 @@ public class MakeMethodConstructorFix implements IntentionAction {
   }
 
   public void invoke(@NotNull Project project, Editor editor, PsiFile file) throws IncorrectOperationException {
+    if (!CodeInsightUtilBase.preparePsiElementForWrite(myMethod)) return;
     myMethod.getReturnTypeElement().delete();
-  }
-
-  public boolean isQuickFix() {
-    return true;
   }
 
   public boolean startInWriteAction() {
